Move ConcatWS to another file
